Serializable Condition using `enum_dispatch`
