+++ crates/control/Cargo.toml
@@ -14,6 +14,7 @@ filtering = { workspace = true }
 framework = { workspace = true }
 kinematics = { workspace = true }
 log = { workspace = true }
+motionfile = { workspace = true }
 nalgebra = { workspace = true }
 ordered-float = { workspace = true }
 projection = { workspace = true }
